线程池基本信息

+++ tomato-components/tomato-component-thread/src/main/java/com/tomato/thread/proxy/RejectedProxyInvocationHandler.java
@@ -6,6 +6,7 @@ import lombok.extern.slf4j.Slf4j;
 import java.lang.reflect.InvocationHandler;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
+import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicLong;
 
 /**
@@ -22,7 +23,7 @@ public class RejectedProxyInvocationHandler implements InvocationHandler {
     /**
      * 记录拒绝策略执行次数
      */
-    private final AtomicLong rejectCount;
+    private final AtomicInteger rejectCount;
 
     @Override
     public Object invoke(Object proxy, Method method, Object[] args) throws Throwable {
